SecurityProtocol and Redesign/Simplify
(iwen65)First redesign to make things easier and more readable. All important params were values that had been stored as properties of the FinTsClient.

If this is connected as close as this it is better to pass the client as ref parameter, that makes the hole method much more flexible and extensible without breaking changes.

ConnectionDetails are a part of the client too. We can handle the new added ServiceProtocolType inside the common connectioninfo without breaking changes too. Mostly it will be TLS12 but who knows.